*** empty log message ***

+++ windjview/libdjvu/GRect.cpp
@@ -275,9 +275,15 @@ operator*(int n, GRectMapper::GRatio r )
      they encountered overflow issues.  Let's use long long ints. */
   llint_t x = (llint_t) n * (llint_t) r.p;
   if (x >= 0)
-    return   ((r.q/2 + x) / r.q);
+//< Changed for WinDjView project
+//    return   ((r.q/2 + x) / r.q);
+    return static_cast<int>((r.q/2 + x) / r.q);
+//>
   else
-    return - ((r.q/2 - x) / r.q);
+//< Changed for WinDjView project
+//    return - ((r.q/2 - x) / r.q);
+    return -static_cast<int>((r.q/2 - x) / r.q);
+//>
 }
 
 inline int 
@@ -286,9 +292,15 @@ operator/(int n, GRectMapper::GRatio r )
   /* [LB] -- See comment in operator*() above. */
   llint_t x = (llint_t) n * (llint_t) r.q;
   if (x >= 0)
-    return   ((r.p/2 + x) / r.p);
+//< Changed for WinDjView project
+//    return   ((r.p/2 + x) / r.p);
+    return static_cast<int>((r.p/2 + x) / r.p);
+//>
   else
-    return - ((r.p/2 - x) / r.p);
+//< Changed for WinDjView project
+//    return - ((r.p/2 - x) / r.p);
+    return -static_cast<int>((r.p/2 - x) / r.p);
+//>
 }
